STEP 4: audit fixes — deprecation, clippy, tombstone horizon, CGNAT guard, coordinator, integration test
Fixes from /tmp/oo comprehensive audit:

- cms.rs: deprecated fetch_update -> try_update (Rust 1.85+ API)
- hll.rs: add Default impl for SubnetHll (clippy::new_without_default)
- detection/lib.rs: remove redundant `as u64` casts on u64 fields
  (unnecessary_cast clippy; u64 * 1000 is already u64)
- aworset.rs: extend purge_expired with 5000ms tombstone horizon
  to prevent distributed resurrection bug (audit Flaw 6):
  premature deletion + delayed gossip delta = re-ban.
  Entries now retained until now_ms.saturating_sub(expires_at)
  >= TOMBSTONE_HORIZON_MS (5s), exceeding max network delay +
  clock skew across the fleet.
- enforcement/coordinator.rs: new CoordinatedEnforcementEngine
  implementing two-phase atomic enforcement (audit Flaw 7):
  Phase 1: WAL append + BPF map insert. On E2BIG (map full),
  gracefully falls back to Tier 1 SHM soft throttle (429) instead
  of failing closed. Phase 2: publish to Seqlock SHM + AWORSet
  CRDT broadcast. Strict CGNAT veto (shared-infra && tier>=3 ->
  tier 2) enforced on both local and remote deltas.
- tests/coherency_invariant_test.rs: integration test pinning
  CGNAT veto contract and two-phase transaction atomicity:
  shared-infra Tier-3 requests downgraded to Tier 2, BPF E2BIG
  falls back to Tier 1, maintenance purges expired bans.

Verification: cargo check --workspace, cargo clippy --workspace
--all-targets -- -D warnings, cargo test --workspace --lib
all pass. Bench harness compiles clean.

+++ crates/ramshield-detection/src/lib.rs
@@ -971,7 +971,7 @@ impl DetectionEngine {
                     if tier != ramshield_cgnat::CGNAT_TIER_ALLOW {
                         self.shm_table.publish_rule(
                             sk as u64,
-                            cfg.detection.subnet_burst_ttl_secs as u64 * 1000,
+                            cfg.detection.subnet_burst_ttl_secs * 1000,
                             tier,
                             0,
                             true,
@@ -980,7 +980,7 @@ impl DetectionEngine {
                     // P2: fleet-fenced gossip.
                     self.mesh_blocklist.record_ban(
                         r.ip,
-                        cfg.detection.subnet_burst_ttl_secs as u64 * 1000,
+                        cfg.detection.subnet_burst_ttl_secs * 1000,
                         tier,
                     );
                     self.metrics
